+++ Tortuga.Chain/Shared/Tests/CommandBuilders/InsertTests.cs
@@ -142,7 +142,7 @@ public class InsertTests : TestBase
 		}
 	}
 
-#if SQL_SERVER_SDS || SQL_SERVER_MDS || SQL_SERVER_OLEDB //SQL Server has problems with CRUD operations that return values on tables with triggers.
+#if SQL_SERVER_MDS || SQL_SERVER_OLEDB //SQL Server has problems with CRUD operations that return values on tables with triggers.
 
 	[DataTestMethod, BasicData(DataSourceGroup.Primary)]
 	public void Insert_Trigger(string dataSourceName, DataSourceType mode)