+++ Tortuga.Chain/Changelog.md
@@ -1,5 +1,11 @@
 ## Version 5.1
 
+### Features
+
+[#518 Preload Tables by Schema](https://github.com/TortugaResearch/Tortuga.Chain/issues/518)
+
+The `DatabaseMetadata.Preload` functions for SQL Server now accept an optional `schemaName` paramters.
+
 ### Technical Debt
 
 Removed obsolete compiler constants.